% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flast}
\alias{flast}
\title{Fast last}
\usage{
flast(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
Returns the last element of a vector, factor, or any other R object
    which has a method for the \code{length} function.
}
\description{
Fast last
}
\examples{
flast(1:5) ## 5
flast(letters) ## "z"
}
