% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in}
\usage{
x \%nin\% y
}
\arguments{
\item{x, y}{vectors or \code{NULL}.
\code{\link[base]{LongVectors}} not supported for \code{y}.}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
Binary operator which returns a logical vector indicating
    if there is \emph{not} a match for \code{x} in \code{y}.
}
\examples{
c(1, 2, 3,'a','b','c') \%!in\% letters
## TRUE TRUE TRUE FALSE FALSE FALSE
}
