% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.length0}
\alias{is.length0}
\alias{is.length1}
\title{Test the length of an R object}
\usage{
is.length0(x = NULL)

is.length1(x = NULL)
}
\arguments{
\item{x}{an R object.}
}
\value{
A length 1 logical vector.
}
\description{
Test the length of an R object
}
\examples{
is.length0() ## TRUE
is.length0(NULL) ## TRUE
is.length0("") ## FALSE
is.length0(NA) ## FALSE
is.length1("") ## TRUE
is.length1(NA) ## TRUE
is.length1(c(NULL, NA)) ## TRUE
is.length1(NULL) ## FALSE
is.length1(1:2) ## FALSE
}
