% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not.null}
\alias{not.null}
\title{Not null}
\usage{
not.null(x)
}
\arguments{
\item{x}{an R object to be tested.}
}
\value{
\code{TRUE} if object is not entirely \code{NULL}, otherwise \code{FALSE}.
}
\description{
Test whether an R object is not \code{NULL}.
}
\examples{
not.null(0) ## TRUE
not.null(NA_character_) ## TRUE
not.null(NULL) ## FALSE
not.null(c(NULL, 0)) ## TRUE
not.null(c(a=NULL, b=NULL)) ## FALSE
}
