% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tabular}
\alias{tabular}
\title{Create a table in Roxygen comments}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{A \code{data.frame} object to print in the documentation.}
}
\value{
A table of sorts.
}
\description{
Print the output of this function to produce
    a table in documentation produced by Roxygen.
}
