% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_dataset.R
\name{track_dataset}
\alias{track_dataset}
\title{Track a dataset}
\usage{
track_dataset(
  df,
  step_name = NA_character_,
  track_log = "data_manipulation_log",
  step_inc = 1
)
}
\arguments{
\item{df}{the data frame to track}

\item{step_name}{What would you like to call the operation you are tracking?}

\item{track_log}{What would you like to name the object that logs the information}

\item{step_inc}{How many steps are you tracking (default is 1).}
}
\value{
\code{TRUE} if object is not entirely \code{NULL}, otherwise \code{FALSE}.
}
\description{
I use this to keep track of the changes to a dataset from multiple operations.
}
